修复报错
