+++ ui/src/components/settings/cluster-management.test.tsx
@@ -3,7 +3,12 @@ import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { MemoryRouter, Route, Routes, useLocation } from 'react-router-dom'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { flexRender, getCoreRowModel, useReactTable } from '@tanstack/react-table'
+import {
+  ColumnDef,
+  flexRender,
+  getCoreRowModel,
+  useReactTable,
+} from '@tanstack/react-table'
 
 const {
   useClusterList,
@@ -91,7 +96,7 @@ vi.mock('../action-table', () => ({
     columns,
   }: {
     data: Array<Record<string, unknown>>
-    columns: any[]
+    columns: ColumnDef<Record<string, unknown>>[]
   }) => {
     const table = useReactTable({
       data,
