+++ ui/src/components/settings/general-management-runtime.ts
@@ -1,0 +1,15 @@
+export function shouldReloadForAnalyticsChange(
+  previousEnableAnalytics: boolean | undefined,
+  nextEnableAnalytics: boolean
+) {
+  return (
+    typeof previousEnableAnalytics === 'boolean' &&
+    previousEnableAnalytics !== nextEnableAnalytics
+  )
+}
+
+export const browserRuntime = {
+  reloadWindow() {
+    window.location.reload()
+  },
+}
